refactor(clawdstrike-policy-event): extract kinds/privacy from edr/receipt/mod.rs
